# Conflicts:
#	windows/src/main/csharp/ch/cyberduck/ui/controller/InfoController.cs


Former-commit-id: 45a524d48bfb93b82f847fda095bac338c47b7d8

+++ ssh/src/main/java/ch/cyberduck/core/sftp/SFTPDeleteFeature.java
@@ -56,7 +56,7 @@ public class SFTPDeleteFeature extends ThreadedDeleteFeature implements Delete {
         // Await and shutdown
         this.await();
         for(Path file : files) {
-            if(file.isDirectory()) {
+            if(file.isDirectory() && !file.isSymbolicLink()) {
                 callback.delete(file);
                 try {
                     session.sftp().removeDir(file.getAbsolute());
